requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==4.2.2
anyascii==0.3.2
appdirs==1.4.4
apturl==0.5.2
args==0.1.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
bcrypt==3.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
Brlapi==0.8.3
cachetools==5.3.2
catalogue==2.0.10
cattrs==23.2.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
cloudpathlib==0.16.0
colorama==0.4.4
command-not-found==0.3
comtypes==1.2.0
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
cryptography==3.4.8
cupshelpers==1.0
customtkinter==5.2.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.7
darkdetect==0.8.0
dateparser==1.1.8
dbus-python==1.2.18
decorator==4.4.2
deepspeed==0.8.3
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.1
dnspython==2.1.0
docopt==0.6.2
duckduckgo-search==2.8.6
duplicity==0.8.21
ecapture==2.0.2
einops==0.7.0
encodec==0.1.1
entrypoints==0.4
exceptiongroup==1.2.0
fasteners==0.14.1
fastjsonschema==2.19.0
ffmpeg==1.4
filelock==3.13.1
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.2
future==0.18.2
g2pkk==0.1.2
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gTTS==2.5.0
gyp==0.1
hangul-romanize==0.1.0
hjson==3.1.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.20.1
idna==3.3
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==4.6.4
imutils==0.5.4
inflect==7.0.0
itsdangerous==2.1.2
jamo==0.4.1
jaraco.context==4.3.0
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonlines==1.2.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_core==5.5.1
keyring==23.5.0
kiwisolver==1.4.5
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
librosa==0.10.0
llvmlite==0.41.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.9
meson==0.61.2
mistune==3.0.2
monotonic==1.6
more-itertools==8.10.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
natsort==8.0.2
nbconvert==5.3.1
nbformat==5.9.2
netifaces==0.11.0
networkx==2.8.8
ninja==1.11.1.1
nltk==3.8.1
num2words==0.5.13
numba==0.58.1
numpy==1.22.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
openai==0.27.4
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.1
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
platformdirs==4.1.0
playsound==1.3.0
pocketsphinx==5.0.3
pooch==1.8.0
preshed==3.0.9
proglog==0.1.10
progressbar==2.5
protobuf==3.20.3
psutil==5.9.7
ptyprocess==0.7.0
pvcheetah==2.0.0
pvporcupine==3.0.1
pvporcupinedemo==3.0.1
pvrecorder==1.2.1
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.11
pycairo==1.20.1
pycaw==20230407
pycparser==2.21
pycups==2.0.1
pydantic==1.9.1
pydeck==0.8.1b0
pydub==0.25.1
pygame==2.5.2
Pygments==2.17.2
PyGObject==3.42.1
pyjokes==0.6.0
PyJWT==2.3.0
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.5.0
pynndescent==0.5.11
pyparsing==2.4.7
pypinyin==0.50.0
pyRFC3339==1.1
pysbd==0.3.4
python-apt==2.4.0+ubuntu2
python-crfsuite==0.9.10
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
pyttsx==1.1
pyttsx3==2.90
pytube==15.0.0
pytz==2022.1
pyxdg==0.27
PyYAML==6.0.1
referencing==0.31.1
regex==2023.10.3
reportlab==3.6.8
requests==2.31.0
requests-cache==1.1.1
requests-oauthlib==1.3.1
resampy==0.4.2
resemble==1.4.2
retry-requests==2.0.0
rich==13.7.0
rotary-embedding-torch==0.3.6
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
scour==0.38.2
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.3.1
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.1
srsly==2.4.8
srt==3.5.3
streamlit==1.21.0
SudachiDict-core==20230927
SudachiPy==0.6.8
sympy==1.12
systemd-python==234
tensorboard==2.15.1
tensorboard-data-server==0.7.2
testpath==0.6.0
thinc==8.2.2
threadpoolctl==3.2.0
tiktoken==0.3.3
tk==0.1.0
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
tornado==6.4
tqdm==4.66.1
trainer==0.0.36
traitlets==5.14.0
transformers==4.36.2
triton==2.1.0
TTS==0.22.0
twilio==8.11.0
typer==0.9.0
typing_extensions==4.8.0
tzlocal==5.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
umap-learn==0.5.5
unattended-upgrades==0.1
Unidecode==1.3.7
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==1.26.5
usb-creator==0.3.7
validators==0.22.0
vboxapi==1.0
vosk==0.3.45
wadllib==1.3.6
wasabi==1.1.2
watchdog==3.0.0
Wave==0.0.2
wavio==0.0.8
weasel==0.3.4
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
wikipedia==1.4.0
wolframalpha==5.0.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xmltodict==0.13.0
yarl==1.9.3
zipp==1.0.0
